Bugs with timing and changing of analog reference fixed and testet

+++ lib/General_Purpose_Functions/General_Purpose_Functions.h
@@ -47,8 +47,8 @@ enum supplyStatusFlag{
 
 typedef struct supplyStatusStruct
 {
-    uint16_t vcc;
-    uint16_t vSupercap;
+    double vcc;
+    double vSupercap;
     enum supplyStatusFlag statusFlag;
 };
